dereference paths inside maps
Right now, you can only reference singe value with refernences, but
for non-scalar values such as lists and maps, there is no way to
actually reference values _within_ those values. So this:

```yaml
let:
  x:
    "y":
      "z": "Hello There"
do x.y.z
```

Will actually evaluate to the map contained at `x`. in other words,
the path is completely ignored.

This allows references to have a "path" which is `.` separated which
will then be used to produce a value. So the above example now
produces "Hello There"