Add multi-arg functions via list destructing of arguments
Sometimes it makes sense to be able to define a function with multiple
arguments like:

```yaml
add(x,y): {+: [x,y]}
```

However, PlatformScript functions only take a single argument.

We don't change this fundamental dynamic. If you want to call a
function that needs multiple values, pass it a list. What we do
provide is a syntax for destructuring arguments so that you can
capture mulitple values from a list inside a function.