Use `$` sigil for all references, including inside strings
Because our current references are just "non-quoted" strings, it means
that there is no way to disabmiguate them from other strings if they
pass through another languge. Imagine, for example, that we parse the
followinrg PlatformScript into JSON

```yaml
- reference
- "reference"
```

This contains a list of two references but when we
parse it into its equilavent JSON, we get:

```json
[
  "reference",
  "reference",
]
```

There is no way to tell which one is the reference and which one is
the string even if we convert it back to YAML. Unfortunately, when it
comes to embedding YAML into various places, this is all to common
that it is parsed into memory, and then re-serialized later,
potentially losing key information.

Our solution to this is to explicitly mark anything that should be a
reference with a `$` sigil.

So this:

```yaml
import:
 - names: [five, id]
   from: "./nodeps.yaml"
 - names: [hello]
   from: "./imports-relative.yaml"

do:
  - id: [five, hello]
```

becomes this:

```yaml
$import:
 - names: [five, id]
   from: "./nodeps.yaml"
 - names: [hello]
   from: "./imports-relative.yaml"

$do:
  - $id: [$five, $hello]
```

The above can be parsed, and re-serialized, and still be executed
correctly, whereas the same is not true of the first.

This ameliorates the problem, but it does not solve it. For example,
what do we do with this situation:

```
- $reference
- '$reference'
```

We should either say: "$x", '$x', and $x all are the same
thing (probably the best idea?), and then have an escaping sequence.

This also includes a spike on string interpolation which wasn't
strictly necessary, but it got intermingled with this change
inextricably while I was trying to determine the ramifications of
mixing strings and references. As a result, each double quoted string,
and unquoted string has a number of "holes" which need to be plugged
before the string can be rendered. Right now, the only thing that can
plug a hole is a reference, and in fact it is the presence of the
references that marks the hole. This brings up the following
questions:

* what is the best syntax for holes. It should have the ability to
embed arbitrary PlatformScript expressions in them and not just
references. Some of the proposals: `$()`, `%()`, `(())`, E.g.

```yaml
Hello %({$capitalize: "world" })
```
* Should there be a separte type for a string template? Right now any
string can have holes, but maybe we should represent it as
`PSTemplate`? This feels more right than what we have now.