Enable publishing to NPM
In order to use PlatformScript inside Browser applications that are
build with Node tooling, as well as being able to use it from within
Node itself, we will need to publish it to NPM.

This uses https://github.com/denoland/dnt in order to build an NPM
package out of your Deno package.

Every time a tag like `platformscript-v1.0.0` is published, it will
build the npm package and push it to npmjs.com. This uses the tag name
as the sole source of truth. If there is a tag, it will be
published. It is up to other logic to decide when to tag and when not
to tag.