Add API for easy embedding
In order to use platform script it needs to be easy to instantiate and
call from within Deno, Browser, and Node. To do this, we need to
present a one-stop api where you can define your global scope, and
then either load modules or execute snippets from within it.

This defines a single interface `PlatformScript` which deals only in
Promise compatible APIs. The globals that you pass into this
environment will be available in all modules and snippets it runs.