Remove ability to define modules as TS/JS
This feature required using dynamic `import()` which threw Webpack for
a a loop, and also was a potential security risk.

We now have the ability to extend the list of global symbols when you
create the PlatformScript interpreter. That is the point at which you
can define any global functions you want, which can then be referenced
by your PS modules. So this feature is not needed and is being removed.