Add template literal syntax based on %(expression) syntax
Working with strings and easily constructing them based on the current
lexical environment is a key feature of PlatformScript. However, we
only had the ability subsititue bare references a-la PERL 5.

```yaml
"hello $x"
```

This adds the ability to add any valid PlatformScript expression
inside a string, but instead of just adding a reference, we surrond
the expression with a `%()` delimiter. The %() was chosen because it
contrasts with `$` and `[]` and `{}` all of which are used to create
PlatformScript primitives.

```yaml
hello %($x)
```

To do this, we introduce a new PlatformScript primitive: `PSTemplate`
which contains its value, as well as a list of embedded expressions,
and then string offesets at which they occur.

When a template is evaluated, it is converted into a `PSString` by
evaluating each expression and plugging it into the string at its
index range.