Embed external values into PlatformScript
The entire point of PlatformScript is to compose pieces of the
platform. In order to do that, we have to be able to represent those
pieces as PlatformScript values that can be passed around.

This introduces a new type `PSExternal` which can appear anywhere a
`PSValue` can. It supports de-referencing via an optional `view`
property which is a function receiving a path that should produce a
new `PSValue`

To help creating values for embedders, there is a set of top-level
constructors for creating PlatformScript values from JavaScript
values. e.g.

```ts
import * as ps from 'platformscript';

let scope = ps.map({
  num: ps.number(10),
  bool: ps.boolean(true),
  extern: ps.external({}),
})

let platformscript = ps.createPlatformScript(scope);

let prog = platformscript.parse("Hello %($num). That is %($bool)");

await platformscript.eval(prog);
```