Recursively evaluate map values
When reducing a tree of PlatformScript you need to be able to define
maps of values that are also themselves reduced, otherwise there is no
way to define code that calls functions that evaluate other functions
by passing pieces of their body to them as arguments.

Now, however, when we evaluate a map, we create a new map that has the
exact same keys as the old map, but the difference is that that values
are evaluated.