Add ability to evaluate a value as a PlatformScript module.
Sometimes we want to be able to treat a snippet of code as a fully
formed module that can depend on other modules and define symbols.

There is a bit of a risk in the sense of what do you do if you have
two different virtual modules that have the same URL? This should
probably be an error.