Ensure function bodies are lexically bound
Houston, we had a problem! functions bodies were not lexically bound😬

That means the following code did not work as expected:

```
$let
  x: world
  say: { $(): "%(hello) %($x)" }
$do:
  {$say: true }
```

Instead of printing "hello world" as expected, it failed with a
ReferenceError of `x` is not defined.

We fix this by using a two-phase computation. Before, we had a single
phase of evaluation (`eval`) to convert a value into a result value:

```
value -eval()-> result
```

Now, we have two phases, `bind` and `eval`

```
value -bind()-> bound -eval()-> result
```

Currently binding a scope to a value, means performing all reference
lookups ahead of time and making sure that we have all requisite
values that we will need for evaluation. So, for example in our
template, its ealuation looks like this:

```
"%(hello) %($x)" -bind()-> "%(hello) %(world)" --eval()--> "hello world"
```

for function invocations, which are represented as maps whose first
key is a `ref`, we dereference the function eagerly and
convert the `map` into a `fncall` value which is a first-class PlatformScript
value now. This can ensure that just like we substitute a reference value,
we can also always make sure that any functions that were in scope
when a block of code is evaluated, will remain in scope. Let's take
the following code as an example:

```yaml
$let:
  salutation: hello
  greet: { $(to): "%($salutation) %($to)" }
  say: { $(): { $greet: world } }
```

in the binding, the map `$greet: world` is converted from a `map`
value into a `fncall` and the function referenced by `greet` is
captured from the scope at that point, so when we eval, we have
already looked up the function and know it is there. In the phased
computation:

```
map($greet: world) -bind()-> fncall(greet,"world") -eval()-> string("hello world")
```

This multi-phase expansion lays the groundwork for a macro system
which we will surely have.

Function parameters
-------------------

There are some breaking changes related to the way functions are
constructed. Specifically, functions _prevent_ their parameters from
being bound during the `bind()` operation. This must be, because in
the function `$(x): hello %($x)` we will not know the value of `x`
until we actually call the function, so binding it too soon would be a
big, fat error.

However, this means that we now need to track the _name_ of a function
parameter in a function so that we can be sure to exclude it from the
values being dereferenced during `bind()`. While this is not strictly
necessary for `native` functions, I went ahead and made naming the
parameter mandatory there because it will help with `intellisense` on
native functions, and also because we will need native functions to
provide type information about their inputs and outputs as well.