fix scoping for native functions
The environment of native functions only contained the lexical scope
that was bound. However, an environment is augmented by a function
parameter when its body is evaluated. We need to make sure that the
environment passed to the native function contains this augmented
scope rather than the just the scope when it was bound as a function
call.