Allow calling fns that are referenced by a fn parameter
When converting a map into a function call, we were requiring that
the function be defined at the time of binding. this works module
imports or global functions, but what if the function reference is
introduced into a scope via a function parameter? We need to leave it
as a reference and then look up the reference when we bind the
function body before evaluation of the function.

This changes the name of the `FnCall -> FnCallContext` and then the
PSValue of type `fncall` records only the fact of the function call
itself, and does not create an environment. This is better in that the
`FNCall` is just data.

Note that this breaks the `$do/$let` statement... as you can now not
introduce a `$let` statement after `$do`. You either need to but
`$let` before, or not have it at all.