Add installer with instructions to home page.
In order to install PS locally, we need an easy way to download
it. Rather than go through the rigamarole of building a custom
executable with all of its current limitations, we can just use the
`deno` installer which will set up the bin script for us. E.g.

```shellsession
$ deno install -Afr https://frontside.com/platformscript/pls
```

> Eventually, we'd like this to be `deno install -Afr https://pls.pub`

This adds a route which redirects to the `pls.ts` executable in the
latest source. That executable has been stripped down to its bare
minimum. We'll build it up from there.

Also, this includes the docs.