Add a simple introduction to PlatformScript
The "placeholder" text put in for the introduction is a broken
window now that the site is live.

fix it by giving a very brief technical overview of the language. Note
that because this is the intro in the docs, it is more technical than
the feature list on the home page, although missing practical concerns
should definitely be on this list.