🐛Fix main layout which was converted to flex
We changed the main layout to be a flex box, but it was still being
marked as `h-screen`. This had the effect of putting the footer over
the main content once you scroll past the fold.

This removes that `h-screen` designation to set the flex box free and
let it flow.