✨Add run command and documentation
As part of the documentation for the install and getting started, we
need to be able to print the version of platform script, and also be
able to run some platform script code. This was suprisingly difficult,
so we want to be able to put in place a system to make adding
commands, specifying their options, and automatically generating help
for them simple.

This uses the "router" strategy where commands are a hierarchy of word
matches, and then the options are like query params for those
matches, so the main dispatch can look like this:

```ts
dispatch(["pls", ...args], {
  "pls": [PlsCommand, {
    "run :MODULE": RunCommand,
    "test FILES...": TestCommand,
  }]
});
```

This lets the parent command `PlsCommand` perform any asynchrony it
wants before finally yielding to its "children" which will allow us to
do any global setup, which again, can be async. It also allows us
to add advanced concurrent global options like a `--watch` parameter
that spawns and restarts its children instead of just running them.

At the root is the `PlsCommand` which defines the `--version` options
and `--help` options. There is a single command below the root (for
now) which is the `RunCommand`

The run command accepts a MODULE from either the local file system or
the network, evaluates that module, and then prints out its value.

The docs have been updated to contain the installation process, and a
way to make sure that your installation is working.