fix playground layout
The playground does not have any implicit height because it is set by
the monaco editor.

This adds the "flex-grow: 1" property to make it expand to fill up all
the room left over in the middle of the flex box. That way, it will
always fill the space between the header and the footer.