new import syntax for modules
Import syntax and module semantics are suboptimal. See
https://github.com/thefrontside/platformscript/issues/57 for a
write-up.

To summarize, this makes imports a map instead of a list:

**before**

```yaml
$import:
  - names: [capitalize,decapitalize]
    from: https://pls.pub/std@1.0.0/strings.yaml
  - names: [parse]
    from https://pls.pub/x/parse.yaml
```

**after**

```yaml
$import:
  capitalize, decapitalize: https://pls.pub/std@1.0.0/strings.yaml
  parse: https://pls.pub/x/parse.yaml
```

This also simplifies the module semantics. Modules do not have a list
of symbols and a value, they just have a value, and that value you can
imported key by key:

```yaml
$import:
  something: https://pls.pub/x/things@1.0.0/mod.yaml

```

Or, use the `<<` merge key after a symbol name to indicate that the
_entire module value_ should be bound to that name.

```yaml
$import:
  everything<<: https://pls.pub/x/things@1.0.0/mod.yaml
```

There is not currently a method to alias imported names, but that should be
provide for eventually.