add ability to ask `pls` what its version is
In order to properly have the `pls --version` command work properly,
we need to actually track internally what the version is.

This adds a `version.json` file which contains the actual version so
that it can be used by the interpreter. JSON was chosen as the format
so that in the event that we want to add tooling to the repository
that is not in JavaScript, it will be easy to read.

The signoff script that generates the "next release" pull request has
been modified to update the version file.