✨Add stdlib to house commond PlatformScript modules
We need a way to have shared, common modules for PlatformScript.

This will mount eveything that is in `/stdlib` at
`https://pls.pub/std`

The `version` parameter is captured in the route because eventually we
will want to use pinned versions of the standard library. However, for
now, this will always use the specific version at the version of the
website.