🐛Fix problem with NPM package generation due to YAML upgrade
We upgrade to use the `yaml` package and not the `yaml-ast-parser`
package. However, we needed to pin the build at a particular version
which apparently caused problems with older versions of DNT.

However, this has been fixed with
https://github.com/denoland/dnt/pull/246

This upgrades the NPM build to use the latest version of DNT, and it
also runs the NPM build as part of the branch verification. That way,
if the NPM build breaks in the future, we'll know about it immediately.